requirements.txt
python-telegram-bot
google-generativeai
pandas
newspaper3k
lxml_html_clean
openpyxl